Update
A partir de aqui comienza la version 2.0.0.
