+++ index.html
@@ -14,7 +14,7 @@
 </head>
 <body class="body--light">
   <main id="main-container" class="main--light">
-    <section id="section-hero-container" class="section-hero-container section-hero-container-light">
+    <section id="section-hero-container" class="section-hero-container--light">
       <header id="header-container">
         <the-logo></the-logo>
       </header>
@@ -28,24 +28,7 @@
         <nav-bar></nav-bar>
       </nav>
       <footer id="footer-container">
-        <div id="footer-content-container-1" class="footer-content-container">Make with 💜 with:</div>
-        <div id="footer-content-container-2" class="footer-content-container">
-          <figure>
-            <img src="./src/assets/vectors/HTML5_Badge.svg" alt="html-logo">
-          </figure>
-          <figure>
-            <img src="./src/assets/png/JavaScript-logo.png" alt="JavaScript-logo">
-          </figure>
-          <figure>
-            <img src="./src/assets/vectors/css-3.svg" alt="css-logo">
-          </figure>
-          <figure>
-            <img src="./src/assets/vectors/Sass_Logo_Color.svg" alt="css-logo">
-          </figure>
-          <figure>
-            <img src="./src/assets/vectors/webcomp.svg" alt="JavaScript-logo">
-          </figure>
-        </div>
+        <footer-content></footer-content>
       </footer>
     </aside>
   </main>
