+++ sass-watch.sh
@@ -19,4 +19,7 @@ THELOGOCSS="${CSS}the-logo.css"
 NAVBARSASS="${SASS}navbar.scss"
 NAVBARCSS="${CSS}navbar.css"
 
-sass --watch $MAINSASS:$MAINCSS $DARKTHEMETOGGLESASS:$DARKTHEMETOGGLECSS $INDEXCARROUSELSASS:$INDEXCARROUSELCSS $THELOGOSASS:$THELOGOCSS $NAVBARSASS:$NAVBARCSS
+FOOTERCONTENTSASS="${SASS}footer-content.scss"
+FOOTERCONTENTCSS="${CSS}footer-content.css"
+
+sass --watch $MAINSASS:$MAINCSS $DARKTHEMETOGGLESASS:$DARKTHEMETOGGLECSS $INDEXCARROUSELSASS:$INDEXCARROUSELCSS $THELOGOSASS:$THELOGOCSS $NAVBARSASS:$NAVBARCSS $FOOTERCONTENTSASS:$FOOTERCONTENTCSS
