--- src/components/index-footer/_index-footer.scss
@@ -1,12 +1,0 @@
-
-
-.footer-content-container {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 20px;
-}
-
-#index-footer figure img {
-  width: 28px;
-}
