+++ src/styles/css/index-carrousel.css
@@ -3,4 +3,14 @@
   padding: 0;
 }
 
+#index-carrousel-wrapper {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: start;
+  justify-content: center;
+  gap: 25px;
+}
+
 /*# sourceMappingURL=index-carrousel.css.map */
