+++ src/styles/css/navbar.css
@@ -6,15 +6,20 @@
 #navbar {
   display: flex;
   flex-direction: column;
-  width: 100%;
+  gap: 28px;
+  padding-inline-end: 40px;
+  font-family: var(--header-font);
+  font-size: 2.8rem;
+  font-weight: 200;
 }
 
 .navbar-link {
   list-style: none;
+  text-align: right;
 }
-
 .navbar-link a {
   text-decoration: none;
+  color: inherit;
 }
 
 /*# sourceMappingURL=navbar.css.map */
