+++ src/styles/css/the-logo.css
@@ -21,11 +21,12 @@
 #logotype h1 {
   font-family: var(--the-font);
   font-size: 6.4rem;
-  font-weight: 400;
+  font-weight: 200;
 }
 #logotype h2 {
   font-family: var(--header-font);
   font-size: 2.8rem;
+  font-weight: 200;
 }
 
-/*# sourceMappingURL=inside.css.map */
+/*# sourceMappingURL=the-logo.css.map */
