+++ src/styles/main.css
@@ -36,13 +36,36 @@ the-logo {
 
 index-carrousel {
   grid-column: 4/9;
-  grid-row: 1/6;
+  grid-row: 1/5;
   width: 100%;
 }
 
 dark-theme-toggle {
+  grid-column: 2/3;
+  grid-row: 1/2;
+  justify-self: end;
   width: 35px;
   height: 35px;
+  margin-block-start: 40px;
+  margin-inline-end: 40px;
+  cursor: pointer;
+}
+dark-theme-toggle:hover {
+  transform: scale(1.25) rotate(340deg);
+}
+
+nav-bar {
+  display: flex;
+  align-items: center;
+  justify-content: end;
+  width: 100%;
+  height: 100%;
+}
+
+footer-content {
+  width: 100%;
+  height: 100%;
+  padding-inline-end: 40px;
 }
 
 * {
@@ -118,7 +141,7 @@ a {
 
 #navbar-container {
   grid-column: 1/3;
-  grid-row: 4/7;
+  grid-row: 3/7;
   width: 100%;
 }
 
@@ -131,6 +154,30 @@ a {
   justify-content: end;
   gap: 15px;
   width: 100%;
+}
+
+.body--light {
+  color: var(--light-theme-foreground-secondary-color);
+}
+
+.body--dark {
+  color: var(--dark-theme-foreground-secondary-color);
+}
+
+.main--light {
+  background-color: var(--light-theme-background-secondary-color);
+}
+
+.main--dark {
+  background-color: var(--dark-theme-background-primary-color);
+}
+
+.section-hero-container--light {
+  background-color: var(--light-theme-background-primary-color);
+}
+
+.section-hero-container--dark {
+  background-color: var(--dark-theme-background-secondary-color);
 }
 
 /*# sourceMappingURL=main.css.map */
