+++ src/styles/main.scss
@@ -7,6 +7,8 @@
 @use '../styles/sass/the-logo-tag.scss';
 @use '../styles/sass/index-carrousel-tag.scss';
 @use '../styles/sass/dark-theme-toggle-tag.scss';
+@use '../styles/sass/nav-bar-tag.scss';
+@use '../styles/sass/footer-content-tag.scss';
 
 * {
   margin: 0;
@@ -81,7 +83,7 @@ a {
 
 #navbar-container {
   grid-column: 1 / 3;
-  grid-row: 4 / 7;
+  grid-row: 3 / 7;
   width: 100%;
 }
 
@@ -94,4 +96,29 @@ a {
   justify-content: end;
   gap: 15px;
   width: 100%;
+}
+
+// Dark mode
+.body--light {
+  color: var(--light-theme-foreground-secondary-color);
+}
+
+.body--dark {
+  color: var(--dark-theme-foreground-secondary-color);
+}
+
+.main--light {
+  background-color: var(--light-theme-background-secondary-color);
+}
+
+.main--dark {
+  background-color: var(--dark-theme-background-primary-color);
+}
+
+.section-hero-container--light {
+  background-color: var(--light-theme-background-primary-color);
+}
+
+.section-hero-container--dark {
+  background-color: var(--dark-theme-background-secondary-color);
 }
