+++ src/styles/sass/dark-theme-toggle.scss
@@ -3,32 +3,17 @@
   padding: 0;
 }
 
-.theme-toggle-icon {
+#theme-toggle-icon {
+  display: inline-block;
   width: 35px;
   height: 35px;
+  background-size: contain;
+}
+
+.dark-theme-toggle-icon {
   background-image: url(../../assets/vectors/dark-theme-icon.svg);
 }
 
-.body--light {
-  color: var(--light-theme-foreground-secondary-color);
-}
-
-.body--dark {
-  color: var(--dark-theme-foreground-secondary-color);
-}
-
-.main--light {
-  background-color: var(--light-theme-background-secondary-color);
-}
-
-.main--dark {
-  background-color: var(--dark-theme-background-primary-color);
-}
-
-.hero-background--light {
-  background-color: var(--light-theme-background-primary-color)
-}
-
-.hero-background--dark {
-  background-color: var(--dark-theme-background-secondary-color);
+.light-theme-toggle-icon {
+  background-image: url(../../assets/vectors/light-theme-icon.svg);
 }
