+++ src/styles/sass/footer-content-tag.scss
@@ -1,0 +1,5 @@
+footer-content {
+  width: 100%;
+  height: 100%;
+  padding-inline-end: 40px;
+}
