+++ src/styles/sass/index-carrousel-tag.scss
@@ -1,5 +1,5 @@
 index-carrousel {
   grid-column: 4 / 9;
-  grid-row: 1 / 6;
+  grid-row: 1 / 5 ;
   width: 100%;
 }
