+++ src/styles/sass/index-carrousel.scss
@@ -2,3 +2,16 @@
   margin: 0;
   padding: 0;
 }
+
+#index-carrousel-wrapper {
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: start;
+  justify-content: center;
+  gap: 25px;
+}
+
+.index-carrousel-paragraph {
+}
