+++ src/styles/sass/nav-bar-tag.scss
@@ -1,0 +1,7 @@
+nav-bar {
+  display: flex;
+  align-items: center;
+  justify-content: end;
+  width: 100%;
+  height: 100%;
+}
