+++ src/styles/sass/navbar.scss
@@ -2,16 +2,22 @@
   margin: 0;
   padding: 0;
 }
+
 #navbar {
   display: flex;
   flex-direction: column;
-  width: 100%;
+  gap: 28px;
+  padding-inline-end: 40px;
+  font-family: var(--header-font);
+  font-size: 2.8rem;
+  font-weight: 200;
 }
 
 .navbar-link {
   list-style: none;
-}
-
-.navbar-link a {
-  text-decoration: none;
+  text-align: right;
+  a {
+    text-decoration: none;
+    color: inherit;
+  }
 }
