+++ src/styles/sass/the-logo.scss
@@ -20,10 +20,11 @@
   h1 {
     font-family: var(--the-font);
     font-size: 6.4rem;
-    font-weight: 400;
+    font-weight: 200;
   }
   h2 {
     font-family: var(--header-font);
     font-size: 2.8rem;
+    font-weight: 200;
   }
 }
